Constant config works. tests separated.
