+++ pallets/constant-config/Cargo.toml
@@ -22,28 +22,28 @@ version = '1.0.6'
 default_features = false
 git = 'https://github.com/paritytech/substrate.git'
 package = 'frame-support'
-rev = '3e651110aa06aa835790df63410a29676243fc54'
+rev = '7c751574d8c53108e78440bfa5dffb80ab644972'
 
 [dependencies.system]
 default_features = false
 git = 'https://github.com/paritytech/substrate.git'
 package = 'frame-system'
-rev = '3e651110aa06aa835790df63410a29676243fc54'
+rev = '7c751574d8c53108e78440bfa5dffb80ab644972'
 
 [dependencies.runtime-primitives]
 default_features = false
 git = 'https://github.com/paritytech/substrate.git'
 package = 'sp-runtime'
-rev = '3e651110aa06aa835790df63410a29676243fc54'
+rev = '7c751574d8c53108e78440bfa5dffb80ab644972'
 
 [dev-dependencies.primitives]
 default_features = false
 git = 'https://github.com/paritytech/substrate.git'
 package = 'sp-core'
-rev = '3e651110aa06aa835790df63410a29676243fc54'
+rev = '7c751574d8c53108e78440bfa5dffb80ab644972'
 
 [dev-dependencies.runtime-io]
 default_features = false
 git = 'https://github.com/paritytech/substrate.git'
 package = 'sp-io'
-rev = '3e651110aa06aa835790df63410a29676243fc54'
+rev = '7c751574d8c53108e78440bfa5dffb80ab644972'
